% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_phecode_levels.R
\name{add_phecode_levels}
\alias{add_phecode_levels}
\title{Add phecode levels to dataframe}
\usage{
add_phecode_levels(
  phecode_data,
  phecode_column = phecode,
  remove_phecode_column = TRUE
)
}
\arguments{
\item{phecode_data}{Dataframe with a column encoding phecodes (make sure
codes are normalized via \code{\link{normalize_phecodes}}).}

\item{phecode_column}{Unquoted name of the column containing phecodes}

\item{remove_phecode_column}{Should the original phecode column be kept in the data?}
}
\value{
Dataframe with phecode levels added as three integer columns
}
\description{
Takes a dataframe with a phecode column and replaces it with the
hierarchichal levels of the phecodes as columns \verb{code_l1, code_l2} and
\code{code_l3}.
}
\examples{

patient_data <- dplyr::tribble(
  ~patient,    ~code, ~counts,
  1, "250.23",      7,
  1, "250.25",      4,
  1, "696.42",      1,
  1, "555.21",      4,
  2, "401.22",      6,
  2, "204.21",      5,
  2, "735.22",      4,
  2, "751.11",      2,
)

# Default removes original phecode column
add_phecode_levels(patient_data, code)

# Can keep original column as well
add_phecode_levels(patient_data, code, remove_phecode_column = FALSE)

}
