% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_colors.R
\name{category_colors}
\alias{category_colors}
\title{PheCode category colors}
\usage{
category_colors(return_df = FALSE)
}
\arguments{
\item{return_df}{Return results as a dataframe with \verb{category`` and }color``
columns? If \code{FALSE} then named array is returned.}
}
\value{
named array or dataframe of category and color
}
\description{
Returns a named array of hex colors corresponding to a well-separated color
scheme for plotting Phecode Categories.
}
\examples{

category_colors()

category_colors(return_df = TRUE)
}
