% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phecode_info.R
\name{get_phecode_info}
\alias{get_phecode_info}
\title{Get phecode information}
\usage{
get_phecode_info(codes, what = "description")
}
\arguments{
\item{codes}{Vector of phecodes}

\item{what}{What info should be returned. Options: \code{'description'},
\code{'category'}, \code{'all'}.}
}
\value{
Vector of \code{what} is description or category, dataframe with phecode,
description, and category if \code{what = 'all'}.
}
\description{
Either grabs a vector of descriptions for passed vector of phecodes, or
vector of categories, or a dataframe with both.
}
\examples{

library(dplyr)

# Simulate some random data associated with phecodes
data_w_phecodes <- sample_n(phecode_descriptions, 100) \%>\%
  select(code = phecode) \%>\%
  mutate(random = rnorm(n()))

# Use function to add a column with description
data_w_phecodes \%>\%
  mutate(descrip = get_phecode_info(code, 'description'))

# Use function to add column with category
data_w_phecodes \%>\%
  mutate(category = get_phecode_info(code, 'category'))

# Use function to get dataframe of category and description that can be to join
# with data
bind_cols(
  data_w_phecodes,
  get_phecode_info(data_w_phecodes$code, 'all')
)

}
