% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_phecode_info.R
\name{join_phecode_info}
\alias{join_phecode_info}
\title{Join data with phecode information}
\usage{
join_phecode_info(data_w_phecode, phecode_column = "phecode")
}
\arguments{
\item{data_w_phecode}{Dataframe with a phecode column.}

\item{phecode_column}{Name of column that contains phecodes}
}
\value{
input dataframe with three columns added to it.
}
\description{
Takes a dataframe with a phecode column in it and adds three new columns
about the phecode to it: \code{description}, \code{category}, and \code{category_number}.
}
\examples{

library(dplyr)

data_w_phecode <- sample_n(phecode_descriptions, 100) \%>\%
  select(phecode) \%>\%
  mutate(random = rnorm(n()))

# Default values
data_w_phecode \%>\%
  join_phecode_info()

# Can change name of phecode column
data_w_phecode \%>\%
  rename(jd_code = phecode) \%>\%
  join_phecode_info('jd_code')

}
