% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_phecodes.R
\name{normalize_phecodes}
\alias{normalize_phecodes}
\title{Normalize a phecode vector to padded character}
\usage{
normalize_phecodes(codes)
}
\arguments{
\item{codes}{A vector of phecodes in whatever format your heart desires. E.g.
\code{8.00}, \code{8}, \code{0008.0}, ...}
}
\value{
A vector of 0-padded phecodes.
}
\description{
Depending on if a phecode was ever converted to a number or not it can show
up in a million different formats. This standardizes everything to a 6 digits
number.
}
\examples{
normalize_phecodes(c(8, 8.1, 9.2, 23.4))
}
