% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{phecode_descriptions}
\alias{phecode_descriptions}
\title{Phecode Descriptions}
\format{A data frame with 1,866 rows and 4 variables:
\describe{
\item{phecode}{Normalized phecode (1.2)}
\item{description}{Description of this specific phecode}
\item{category}{What broader category phecode resides in}
\item{category_number}{Integer mapping for category}
\item{phecode_index}{Integer mapping for all phecodes}
}}
\usage{
phecode_descriptions
}
\description{
A dataset containing the info on phecodes in terms of description, category, and index
}
\keyword{datasets}
