% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{phecode_to_levels}
\alias{phecode_to_levels}
\title{Phecode to hierarchy levels}
\format{A data frame with 1,866 rows and 4 variables:
\describe{
\item{phecode}{Normalized phecode (1.2)}
\item{code_l1}{Level one of hierarchy (aka number before decimal point)}
\item{code_l2}{Level two of hierarchy (aka first number after decimal point)}
\item{code_l3}{Level three of hierarchy (aka second number after decimal point)}
}}
\usage{
phecode_to_levels
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
