% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_phecode_pairs.R
\name{rollup_phecode_pairs}
\alias{rollup_phecode_pairs}
\title{Rollup binary phecode pairs}
\usage{
rollup_phecode_pairs(
  patient_code_counts,
  patient_col = patient,
  phecode_col = phecode
)
}
\arguments{
\item{patient_code_counts}{Data containing patient ids, phecodes, and counts
of those phecodes.}

\item{patient_col}{Unquoted name of the column containing patient ids in
passed \code{patient_code_counts}'s dataframe.}

\item{phecode_col}{Unquoted name of the column containing phecodes in passed
dataframe.}
}
\value{
Dataframe in same format as passed \code{patient_code_counts} with code
counts rolled up. This will most likely be longer than the original dataset
because of the addition of potentially previously ommitted codes.
}
\description{
Takes phecodes that have been coded without rollup - E.g. existance of code
008.10 does not imply the existance of 008.00 - and produces a rolledup
version of the dataset. Works with plain patient-phecode pairs and also with
patient-code-count triplets.
}
\examples{

library(dplyr)

patient_data <- tribble(
  ~patient,    ~code,
         1, "250.23",
         1, "250.25",
         1, "696.42",
         1, "555.21",
         2, "401.22",
         2, "204.21",
         2, "735.22",
         2, "751.11",
)


# Also works with just binary occurance pairs
patient_data \%>\%
  rollup_phecode_pairs(patient_col = patient,
                  phecode_col = code)

}
