% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_phecode.R
\name{scale_color_phecode}
\alias{scale_color_phecode}
\title{Phecode color scale for ggplot2}
\usage{
scale_color_phecode()
}
\value{
ggplot2 color scale using \code{category_colors()} colors for each category.
}
\description{
Phecode color scale for ggplot2
}
\examples{

library(dplyr)
library(ggplot2)

sample_n(phecode_descriptions, 300) \%>\%
  left_join(category_colors(return_df = TRUE) \%>\%
              mutate(bias = rnorm(n(), sd = 3)),
            by = 'category') \%>\%
  mutate(val = rnorm(n(), mean = bias)) \%>\%
  ggplot(aes(x = phecode, y = val, color = category)) +
  geom_point() +
  scale_color_phecode() +
  theme(axis.ticks.x = element_blank(),
        axis.text.x = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

}
