% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spare_matrix_to_tidy_df.R
\name{sparse_matrix_to_tidy}
\alias{sparse_matrix_to_tidy}
\title{Tidy a sparse matrix}
\usage{
sparse_matrix_to_tidy(
  sparse_mat,
  rows_id = "x",
  cols_id = "y",
  values_id = "value",
  keep_values = TRUE
)
}
\arguments{
\item{sparse_mat}{Sparse matrix of class \code{dgCMatrix} or its variants}

\item{rows_id}{What is stored in the rows of matrix}

\item{cols_id}{What is stored in the columns on matrix}

\item{values_id}{What the values in the cells are}

\item{keep_values}{Do you want to keep the values of the cells or is the matrix a binary one where just existance matters?}
}
\value{
Tidy df with a row for each non-empty entry in the sparse matrix. If \code{keep_values == TRUE} then there will be a third column containing the value of cell in matrix at pairing.
}
\description{
Tidy a sparse matrix
}
