% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_phewas.R
\name{theme_phewas}
\alias{theme_phewas}
\title{Phewas friendly theming for ggplot2}
\usage{
theme_phewas(phecode_on_x_axis = TRUE)
}
\arguments{
\item{phecode_on_x_axis}{Is the phecode in the plot on the x axis?}
}
\value{
Ggplot2 theme object
}
\description{
Turns off the plots attempt to write out each individual phecode name and
ticks so plot renders faster and cleaner.
}
\examples{
library(dplyr)
library(ggplot2)

data <- sample_n(phecode_descriptions, 300) \%>\%
  left_join(category_colors(return_df = TRUE) \%>\%
              mutate(bias = rnorm(n(), sd = 3)),
            by = 'category') \%>\%
  mutate(val = rnorm(n(), mean = bias))

ggplot(data, aes(x = phecode, y = val, color = category)) +
  geom_point() +
  scale_color_phecode() +
  theme_phewas()


ggplot(data, aes(y = phecode, x = val, color = category)) +
  geom_point() +
  scale_color_phecode() +
  theme_phewas(phecode_on_x_axis = FALSE)

}
