% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_phecode_info.R
\name{join_phecode_info}
\alias{join_phecode_info}
\title{Join data with phecode information}
\usage{
join_phecode_info(
  data_w_phecode,
  phecode_column = phecode,
  cols_to_join = c("description", "category", "category_number", "phecode_index")
)
}
\arguments{
\item{data_w_phecode}{Dataframe with a phecode column.}

\item{phecode_column}{Unquoted column of passed data that contains phecodes}

\item{cols_to_join}{Which columns of phecode info do you want appended to
dataframe?}
}
\value{
input dataframe with three columns added to it.
}
\description{
Takes a dataframe with a phecode column in it and desired new phecode
information columns to it: \code{description}, \code{category}, \code{category_number}, and
\code{phecode_index} (the relative position in ordered phenome of phecode).
}
\examples{

library(dplyr)

data_w_phecode <- sample_n(phecode_descriptions, 100) \%>\%
  select(phecode) \%>\%
  mutate(random = rnorm(n()))

# Default values
data_w_phecode \%>\%
  join_phecode_info()

# Can change name of phecode column
data_w_phecode \%>\%
  rename(jd_code = phecode) \%>\%
  join_phecode_info(jd_code)

# Can choose which columns are added
data_w_phecode \%>\%
  join_phecode_info(cols_to_join = c("description", "phecode_index"))

}
\seealso{
get_phecode_info
}
